Loaded Players with images
Loaded players form selected team with images, names and numbers

+++ Lib/Model/Player.cs
@@ -19,7 +19,7 @@ namespace Lib.Model
         public bool Captain { get; set; }
 
         [JsonProperty("shirt_number")]
-        public long ShirtNumber { get; set; }
+        public int ShirtNumber { get; set; }
 
         [JsonProperty("position")]
         public string Position { get; set; }
